Simplify predicates in `PushDownFilter` optimizer rule (#16362)
* Simplify predicates in filter

* add slt test

* Use BtreeMap to make tests stable

* process edge coner

* add doc for simplify_predicates.rs

* add as_literal to make code neat

* reorgnize file

* reduce clone call
