+++ datafusion/optimizer/src/lib.rs
@@ -61,7 +61,6 @@ pub mod push_down_limit;
 pub mod replace_distinct_aggregate;
 pub mod scalar_subquery_to_join;
 pub mod simplify_expressions;
-mod simplify_predicates;
 pub mod single_distinct_to_groupby;
 pub mod utils;
