+++ datafusion/optimizer/src/push_down_filter.rs
@@ -41,7 +41,7 @@ use datafusion_expr::{
 
 use crate::optimizer::ApplyOrder;
 use crate::utils::{has_all_column_refs, is_restrict_null_predicate};
-use crate::{simplify_predicates::simplify_predicates, OptimizerConfig, OptimizerRule};
+use crate::{simplify_expressions::simplify_predicates, OptimizerConfig, OptimizerRule};
 
 /// Optimizer rule for pushing (moving) filter expressions down in a plan so
 /// they are applied as early as possible.
@@ -783,7 +783,9 @@ impl OptimizerRule for PushDownFilter {
         let new_predicates = simplify_predicates(predicate)?;
         if old_predicate_len != new_predicates.len() {
             let Some(new_predicate) = conjunction(new_predicates) else {
-                return plan_err!("at least one expression exists");
+                // new_predicates is empty - remove the filter entirely
+                // Return the child plan without the filter
+                return Ok(Transformed::yes(Arc::unwrap_or_clone(filter.input)));
             };
             filter.predicate = new_predicate;
         }
